Adding pipe transforming paragraph and input values.
Takes in text to be repeated and user's input. Slices off 32 characters,
finds index whitespace for a good breaking point, slices off both
paragraph and user input to that index to create text : input pairs of
same length. Checks every letter in user input against the sentence he
is supposed to repeat to check for errors and returns html element with
class and letter as string to be used in [innerHTML] binding to
render it.